Creating a MainButtonStyle rather than using a custom view
